pyproject.toml
[project]
name = "vector-database"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "google-genai>=1.15.0",
    "pydantic>=2.11.4",
    "pymilvus[model]>=2.5.8",
    "pypdf2>=3.0.1",
    "requests>=2.32.3",
    "uvicorn>=0.34.2",
]
